\name{twdetect}
\alias{twdetect}
\title{twdetect}

\description{Takes a vector of strings, and a regular expression,
and returns a list of booleans for each potential detect.}

\usage{
twdetect(v, regex)
}

\arguments{
  \item{v}{the string vector}
  \item{regex}{a regular expression.}
}

\value{returns the vector of detection events}

\author{David L Gibbs}





