\name{twdetect_all}
\alias{twdetect_all}
\title{twdetect_all}

\description{Takes a vector of strings, and a regular expression,
and returns a list of booleans for each potential detect_all.}

\usage{
twdetect_all(v, regex)
}

\arguments{
  \item{v}{the string vector}
  \item{regex}{a regular expression.}
}

\value{returns the vector of detect_all events}

\author{David L Gibbs}





