\name{twine-package}
\alias{twine-package}
\docType{package}
\title{fast string operations on a character vector}
\description{fast string operations}
\details{
\tabular{ll}{Package: \tab twine\cr
Type: \tab Package\cr
Title: \tab fast string operations on a character vector\cr
Version: \tab 0.1\cr
Date: \tab 2012-03-04\cr
Author: \tab David L Gibbs\cr
Maintainer: \tab David L Gibbs <gibbsdavidl@gmail.com>\cr
License: \tab GPL (>=2)\cr
Depends: \tab Rcpp (>= 0.9.10)\cr
LinkingTo: \tab Rcpp\cr}
}
\author{David L Gibbs}

\keyword{ package }


