\name{twlocate}
\alias{twlocate}
\title{twlocate}

\description{Takes a vector of strings, and a regular expression,
and returns a list of ints for each potential location.}

\usage{
twlocate(v, regex)
}

\arguments{
  \item{v}{the string vector}
  \item{regex}{a regular expression.}
}

\value{returns the vector of locate events}

\author{David L Gibbs}





