\name{twsplit}
\alias{twsplit}
\title{twsplit}

\description{Takes a vector of strings, and performs string splitting.}

\usage{
twsplit(v, pattern)
}

\arguments{
  \item{v}{the string vector}
  \item{pattern}{the regular expression}
}

\value{returns the split strings}

\author{David L Gibbs}





