\name{twtoken}
\alias{twtoken}
\title{twtoken}
\description{Takes a vector of strings, and a string to split by,
and returns a list of character vectors.}
\usage{
twtoken(v,
s)
}
\arguments{
  \item{v}{the string vector}
  \item{s}{the elements to split by.}
}

\value{returns the vector of string lengths}

\author{David L Gibbs}





