\name{twmatch}
\alias{twmatch}
\title{twmatch}

\description{Takes a vector of strings, and a regular expression,
and returns a list of booleans for each potential match.}

\usage{
twmatch(v, regex)
}

\arguments{
  \item{v}{the string vector}
  \item{regex}{a regular expression.}
}

\value{returns the vector of string lengths}

\author{David L Gibbs}





