\name{twmatch}
\alias{twmatch}
\title{twmatch}
\description{Takes a vector of strings, and a regular expression,
and returns a list of booleans for each potential match.}
\usage{
twmatch(v,
regex,
full = FALSE)
}
\arguments{
  \item{v}{the string vector}
  \item{regex}{a regular expression.}
  \item{full}{full match (T/F)?  {F == partial match}.}
}

\value{returns the vector of string lengths}

\author{David L Gibbs}





